requirements.txt
APScheduler==3.6.3
autopep8==1.7.0
cachelib==0.9.0
cachetools==4.2.2
certifi==2022.6.15
click==8.1.3
cs50==9.2.2
distlib==0.3.5
filelock==3.7.1
Flask==2.2.2
Flask-Session==0.4.0
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.30
numpy==1.23.2
pandas==1.4.4
platformdirs==2.5.2
protobuf==3.20.1
pycodestyle==2.9.1
python-dateutil==2.8.2
python-telegram-bot==13.13
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
SQLAlchemy==1.4.41
sqlparse==0.4.2
termcolor==1.1.0
toml==0.10.2
tornado==6.2
tzdata==2022.1
tzlocal==4.2
virtualenv==20.15.1
Werkzeug==2.2.2
zipp==3.8.1
